metro: do not override process.env completely in prelude
Summary: This changeset tweaks the prelude so as to avoid overriding `process.env` if it already exists, and add tests to verify this behavior. This is useful when a bundle is loaded in a context that already provides a global `process` somehow, ex. Electron/Node.js. We still do want to override `NODE_ENV` in that case so as to have consistency with constant inlining.

Reviewed By: rafeca

Differential Revision: D6702441

fbshipit-source-id: 69dd9ba2303a43db151cbe1877f01e38d45b05b9