Use always the same prologue when minifying (g, r, m, e, d)
Summary:
When minfiying modules in separate threads, global variable count varies from one module to another. This can cause prologues to differ when minified, despite being the same variable. This has empirically proven to be a GZip deoptimizator.

```
__d(function(t,s,e,o,r){'use strict';   ==>  _d(function(g,r,m,e,d){'use strict';
__d(function(t,e,r,s,n){'use strict';   ==>  _d(function(g,r,m,e,d){'use strict';
__d(function(t,s,e,o,r){'use strict';   ==>  _d(function(g,r,m,e,d){'use strict';
```

The diff sets up a minification for all variables to their first lowercase letter, then ensures UglifyJS will not mangle it. This process is **ONLY** done if minification is enabled.

Reviewed By: rafeca

Differential Revision: D8400952

fbshipit-source-id: 3630e791206b28b78a91e5d528de53b17af2ebd5