Make Metro bundles compatible with ES6 Module runtimes
Summary:
ES6 Modules are run with an `undefined` context, and they don't have a `global` binding, but just the standard `window`, so currently Metro bundles cannot run on them.

This should not affect performance, since the new check is only applied to polyfills (that's in total 11 times for the whole bundle in RN).

This diff fixes https://github.com/facebook/metro/issues/239

Reviewed By: mjesun

Differential Revision: D9748662

fbshipit-source-id: ae93c3a68c8a84131e100252019ef7c3e3370648