make metro visualizer an optional dependency (#246)
Summary:
<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

**Summary**

This makes `metro-visualizer` an optional package as currently it's quite heavy on dependencies and takes up a majority percentage of React Native's package install size;

![image](https://user-images.githubusercontent.com/5347038/45516977-abd23600-b7a4-11e8-9bd2-2c677afaf855.png)

And: https://twitter.com/jamonholmgren/status/1040260845857656833?s=21

This will try require the package if the config option is enabled and warn if the visualizer package could not be found but gracefully continue. This means its now up to user-land code whether to include this package in their project.

**Test plan**

All tests pass locally:

![image](https://user-images.githubusercontent.com/5347038/45517195-403c9880-b7a5-11e8-9f49-e133e6cd0146.png)
Pull Request resolved: https://github.com/facebook/metro/pull/246

Differential Revision: D9817764

Pulled By: rafeca

fbshipit-source-id: 81635d75aa4cebdcd332d241c8223c616e74724f