metro: keep track of significant dependencies ordering
Summary:
This fixes what I believe to be a long-time bug in the "delta bundler". It appears the delta bundler doesn't care about the dependencies ordering, because it puts them in a Map, indexed by dependency path. The problem is, the order is actually meaningful as it matches the order of the "dependencyMap" generated by the `collectDependencies` module.

Still, using a `Map` worked most of the time because `Map` keep ordering, just like objects. However, if the dependencies of a file are changed, notably some are removed and some are added, the logic in `traverseDependencies.js` would not keep the correct ordering anymore, as it would delete in the middle but append new modules at the end.

This changeset adds a new field in the delta bundler's `Module` representation to keep track of the list of absolute file paths in the order of the `dependencyMap`. This order just needs to pass through all the way to the module JS wrapper logic.

Reviewed By: mjesun

Differential Revision: D14024934

fbshipit-source-id: c8fc232d321fd97f7a2b5ba0d3d4a2c250d85a6c