Make line and column numbering conventions configurable
Summary:
Adds flags to `metro-symbolicate` that let users control the numbering convention for line and column in the input and output (separately).

The default is 1-based lines and 0-based columns for both input and output (as in the underlying `source-map` library), but we sometimes need to interoperate with tools that follow different conventions. For instance, Node's stack traces are 1:1-based.

Reviewed By: cwdick

Differential Revision: D15279488

fbshipit-source-id: 6c57b568df3f4a33f16f40fa48273bd986a76653