Fix JSON.parse error when read cache files which havn't been closed (#421)
Summary:
**Summary**

Since we build a batch of  `jsbundle` concurrently, occasionally when one process is writing a new file on the file system, at this moment another process try to read the file, it will got an empty file and cause the `JSON.parse` throw `SyntaxError`.

**Test plan**

Added a test.
Pull Request resolved: https://github.com/facebook/metro/pull/421

Differential Revision: D16646846

Pulled By: cpojer

fbshipit-source-id: 2fe97e105beace40de1217c11d787e714249a80e