Change transform worker from a class to an object interface
Summary:
There is no reason for `JSTransformer/worker.js` to export a class as it doesn't have state. This is likely a leftover of the early Metro design. This diff changes the module to export two functions: `transform` and `getCacheKey`, very similar to how Jest does it.

This is in preparation for changing the worker output types to support HBC.

Reviewed By: motiz88

Differential Revision: D20952734

fbshipit-source-id: e848395929d762815f5138fa7a1adfc999494c96