Create `metro-transform-worker` package
Summary:
This diff brings all the worker related functionality into a new package to separate concerns.

Note: `ModuleGraph` also needs to be split out of `metro` at some point to get rid of the circular dependency. This is not a problem given that we lock all `metro-*` packages to the same exact version but definitely not ideal. I will follow-up but it isn't blocking me.

Reviewed By: motiz88

Differential Revision: D21014084

fbshipit-source-id: e25f099d2cb5427627bc1666d590bd2d29eefb22