Increase perceived bundling speed
Summary:
This diff increased perceived bundle speed by reducing the amount of progress events sent to the client.

We do this by not sending events that don't meaningfully change the percentage, and not sending events that decrease percentage.

This results in a more fluid, progressive loading experience that even ends _before_ the terminal reporter instead of significantly after it.

Reviewed By: cpojer

Differential Revision: D21222951

fbshipit-source-id: 7369cdc1b9a9cca4949bc7f562976058b0382855