Add rewriteRequestUrl option
Summary: Adds `config.server.rewriteRequestUrl` to Metro. `rewriteRequestUrl` is a function that will get the chance to rewrite a request or bundle URL before it reaches Metro. This includes the URLs of all incoming HTTP requests, as well as bundle URLs in `/symbolicate` request payloads and in the hot reloading protocol.

Reviewed By: cpojer

Differential Revision: D21455962

fbshipit-source-id: 5f80d4e2edd2d974b6133ffffbad08352f7288ed