Create `@fb-config/jest` and `@fb-tools/jest-transformer`
Summary:
This diff splits the Jest setup into a config and a transformer package, and makes some cleanups like finally turning the Jest config into a JS file that can be shared instead of having two copies that we try to keep in sync (fwiw I don't think we've been running tests with the Android platform for years…). The reason for the split is because the transformer depends on our transform pipeline and Babel and needs to be deployed separately in the future (and will be versioned with Metro's transform pipeline, like it already basically is). Other configuration can be changed without triggering a deploy of tooling, like excluding certain folders etc.

The Jest Sandcastle setup will soon be replaced with a new one by the www DevX team so I'm moving files specific to that setup there. In the past that folder was deployed to www so it made sense to separate the config but that is no longer the case.

NOTE: Please excuse the relative require to the Babel transformer. This is fixed in the next diff, D21715897, which adds `fb-tools/transformer`.

(Note: this ignores all push blocking failures!)

Reviewed By: motiz88

Differential Revision: D21713976

fbshipit-source-id: 36ff4969a852a6c4b4bbf9b833829968f24d70e0