Add option to generate V8 CPU profiles of Buck worker tools
Summary:
Adds a CPU profiling option to `buck-worker-tool`. When the `JS_WORKER_TOOL_CPU_PROFILE` environment variable is set, we will now save a V8 `.cpuprofile` file in the current directory when the worker terminates.

Ordinarily, passing `--cpu-prof` to Node would be sufficient, but it seems that Buck terminates the worker before Node has a chance to write the profile to disk.

Reviewed By: cpojer

Differential Revision: D22023324

fbshipit-source-id: 84f75e2bd39e19b818254cf058749064a463b0a9