Use dependency-injection for the `cacheStores` config option
Summary:
For Metronome (https://fb.quip.com/S2cZA1mV1jLe) the goal is to ship tools as binaries but still allow configuration (`xplat/js/config`) to be changed without triggering a deploy of the tool, as some config changes frequently. This works for most configuration as they are just declaratively defining strings or booleans. The only configuration option I could find that is different is the `cacheStores` option in Metro which consists of an array of class instances either from `metro-cache`'s exported classes, custom sub-classes or completely custom caches.

Once we deploy Metro as a binary, regular users will not have the ability to directly require `metro-cache` from a configuration module (without installing node_modules etc.). Therefore in this diff I am changing the `cacheStores` option to either be the existing array (no breaking change) or be a function that receives the `metro-cache` module via dependency injection.

I also fixed the type definitions in `scripts/dependency-analysis` which should be `ConfigT`, not `InputConfigT`.

Reviewed By: rubennorte

Differential Revision: D22041381

fbshipit-source-id: 941aeaad740a10a0d9ae56afaf592a57ede57477