Type Path/Node as objects
Summary:
Path and node are currently typed as any. That is problematic because it disables the flow checks for these types completely. E.g. Flow doesn't catch the error if a function returns `null` but the return type annotation requires a non-nullable `Node`.

This diff changes the type of `Path` and `Node` to objects with arbitrary keys. This enables flow to type check if the return value has the correct `nullability` and if it is an `object`.

Reviewed By: cpojer

Differential Revision: D23672213

fbshipit-source-id: 5a416f97438b0d8ac1a508cebb255cbe3c559802