Don't traverse the entire AST
Summary:
`generateImportNames` 'traverses' full AST even though it's only interested in visiting the Program Node  (actually, the AST is visited twice: once for building the scope and once for traversing). .

Babel tries to keep the overhead low by not constructing `NodePath`s for nodes that are not actually past to a visitor function but it still has to traverse the full AST in case the visitor is interested in any child node.

This diff adds a `path.stop` after `Program` to avoid a second traversal of the full AST.

Reviewed By: motiz88

Differential Revision: D23707445

fbshipit-source-id: 77465c0628a3a5289eff851e5086da26cf98cefc