Remove JSON from default asset types (#593)
Summary:
**Summary**

JSON files are currently considered both source and assets so they are duplicated in the js bundle and in the assets copied in the final app. I think this is a better default and if someone relies on json files being also included in the bundle they can add it back in their metro config.

Fixes https://github.com/facebook/metro/issues/487

**Test plan**

Test that json files are no longer present in my app assets and the app still works.

Pull Request resolved: https://github.com/facebook/metro/pull/593

Reviewed By: MichaReiser

Differential Revision: D24904641

Pulled By: cpojer

fbshipit-source-id: 934b31fc022b3c32f57c88d17eaee22a2808728a