Haste Regex Perf Improvement
Summary:
Finally, a quick and easy perf win!

Some RegExp in the haste map used `.*someText` which has the issue that `.*` matches everything and the engine has to back track if the remaining string doesn't match `someText`.

The `.*` at the beginning or end are not needed since the RegExps don't use `^` or `$` to force that the string is at the beginning or the end. I simply removed it and wow does that make a difference.

https://pxl.cl/1s5zp

Results for building `MarketplaceHome.entrypoint` on my on-demand with `--reset-cache` using a local metro metronome build.

|        | Run 1  | Run 2  | Run 3  |
|--------|--------|--------|--------|
| Before | 1m 11s | 1m 12s | 1m 12s |
| After  | 1m     | 1m 1s  | 1m 1s  |

Reviewed By: cpojer

Differential Revision: D25180387

fbshipit-source-id: 624635587da2a43c7977761552770800eaa9c667