Support preserving specific names in normalizePseudoGlobals
Summary:
`normalizePseudoGlobals` is an AST pass that Metro's transformer runs before minification, and is conceptually **part of the minification pass**. Specifically, it renames the module wrapper function's parameters according to a consistent scheme, and protects the new names from further mangling by the full minifier. This results in more consistent function prologues and makes the bundle more compressible.

Here, we make it possible to pass a list of "reserved" parameter names which  `normalizePseudoGlobals` will not rename.

Reviewed By: MichaReiser

Differential Revision: D28120786

fbshipit-source-id: 68965757ab7a50c30434526e0c835d428b7fd5a0