Update inspectorProxy with IPV6 localhost address (#662)
Summary:
**Summary**

For IPV4 the webSocketDebuggerUrl defaults to `ws://localhost:8081` but for IPV6 this defaults to `ws://[::]:8081`. This causes issues since `ws://[::]` is not a valid websocket URL, I am guessing it should be `ws:[::1]` which is the IPV6 equivalent.

**Test plan**

The webSocketDebuggerUrl should be `ws:[::1]` rather than `ws://[::]`.

This is related to this issue in Flipper https://github.com/facebook/flipper/issues/1520.

Pull Request resolved: https://github.com/facebook/metro/pull/662

Reviewed By: GijsWeterings

Differential Revision: D29210395

Pulled By: motiz88

fbshipit-source-id: 1522265f786765156aa650c704ad5d2c0605ca24