Added named groups babel transform to rn babel preset (#791)
Summary:
## Summary

Hermes currently does not support regex named groups. Whenever a library is included that uses named groups the rn build with hermes will fail (as it doesn't compile). This is already the case when using jest 27 as this includes a dependency (ansi-styles) that uses named-groups which is transitively included via pretty-format in HMRClient.

Pull Request resolved: https://github.com/facebook/metro/pull/791

Test Plan: no test plan

Reviewed By: arushikesarwani94

Differential Revision: D35359562

Pulled By: robhogan

fbshipit-source-id: 33d5c4c1a6ddf8c916950d39e662f47dfdfc4d86