Reject runServer promise if the HTTP server can't start listening
Summary:
Changelog:

* **Breaking:** Reject `runServer` promise if the HTTP server can't start listening.

Previously the only way to handle an error at this point (e.g. when the requested port is busy) would have been through an `onError` callback and not through the call stack. Here we keep `onError` working but also reject the `runServer` promise if it is still pending. This is a breaking change for integrators who rely on `runServer` resolving "succesfully" when such an error occurs.

Reviewed By: robhogan

Differential Revision: D35849395

fbshipit-source-id: aecc8326db34a219d2d88506379cb7c00f55dfd4